back-end/src/services/packageService.ts
// src/services/packageService.ts
import { DynamoDBService, dynamoDBService } from './dynamoDBService';
import { log } from '../logger';
import { PackageTableItem, PackageVersionTableItem } from '../types';

export class PackageService {
  private db: DynamoDBService;

  constructor() {
    this.db = dynamoDBService;
  }

  public async getPackageByName(name: string): Promise<PackageTableItem | null> {
    try {
      return null; // Stub implementation
    } catch (error) {
      log.error('Error getting package by name:', error);
      throw error;
    }
  }

  public async getPackageById(id: string): Promise<PackageTableItem | null> {
    try {
      return null; // Stub implementation
    } catch (error) {
      log.error('Error getting package by id:', error);
      throw error;
    }
  }

  public async getPackageVersion(packageId: string, version: string): Promise<PackageVersionTableItem | null> {
    try {
      return null; // Stub implementation
    } catch (error) {
      log.error('Error getting package version:', error);
      throw error;
    }
  }

  public async listPackages(): Promise<PackageTableItem[]> {
    try {
      return []; // Stub implementation
    } catch (error) {
      log.error('Error listing packages:', error);
      throw error;
    }
  }

  public async listPackageVersions(packageId: string): Promise<PackageVersionTableItem[]> {
    try {
      return []; // Stub implementation
    } catch (error) {
      log.error('Error listing package versions:', error);
      throw error;
    }
  }
}