back-end/test/packageService.test.ts
import { PackageService } from '../src/services/packageService';
import { dynamoDBService } from '../src/services/dynamoDBService';

// Mock DynamoDB service
jest.mock('../src/services/dynamoDBService', () => ({
    dynamoDBService: {
        getPackageByName: jest.fn(),
        getPackageById: jest.fn(),
        getPackageVersion: jest.fn(),
        listPackages: jest.fn(),
        listPackageVersions: jest.fn()
    }
}));

describe('PackageService', () => {
    let packageService: PackageService;

    beforeEach(() => {
        packageService = new PackageService();
        jest.clearAllMocks();
    });

    // Basic test to ensure service instantiates
    it('should create an instance of PackageService', () => {
        expect(packageService).toBeInstanceOf(PackageService);
    });

    /* Commented out for now - implement when needed
    describe('getPackageByName', () => {
        it('should return null (stub implementation)', async () => {
            const result = await packageService.getPackageByName('test-package');
            expect(result).toBeNull();
        });
    });

    describe('getPackageById', () => {
        it('should return null (stub implementation)', async () => {
            const result = await packageService.getPackageById('test-id');
            expect(result).toBeNull();
        });
    });

    describe('getPackageVersion', () => {
        it('should return null (stub implementation)', async () => {
            const result = await packageService.getPackageVersion('test-id', '1.0.0');
            expect(result).toBeNull();
        });
    });

    describe('listPackages', () => {
        it('should return empty array (stub implementation)', async () => {
            const result = await packageService.listPackages();
            expect(result).toEqual([]);
        });
    });

    describe('listPackageVersions', () => {
        it('should return empty array (stub implementation)', async () => {
            const result = await packageService.listPackageVersions('test-id');
            expect(result).toEqual([]);
        });
    });
    */
});